fix some warning.

+++ src/common/timer/timer_1ms.cpp
@@ -1,3 +1,8 @@
+// Use of this source code is governed by a BSD 3-Clause License
+// that can be found in the LICENSE file.
+
+// Author: caozhiyi (caozhiyi5@gmail.com)
+
 #include "timer_1ms.h"
 
 namespace quicx {
@@ -112,7 +117,7 @@ uint32_t Timer1ms::TimerRun(uint32_t time) {
                 if (next_index >= __timer_capacity) {
                     next_index = next_index % __timer_capacity;
                 }
-                AddTimerByIndex(ptr, next_index);
+                AddTimerByIndex(ptr, uint8_t(next_index));
             }
         }
     }
