+++ src/common/timer/timer_container.cpp
@@ -1,3 +1,8 @@
+// Use of this source code is governed by a BSD 3-Clause License
+// that can be found in the LICENSE file.
+
+// Author: caozhiyi (caozhiyi5@gmail.com)
+
 #include <assert.h>
 #include <algorithm> // for min
 
@@ -20,11 +25,11 @@ TimerContainer::~TimerContainer() {
 }
 
 bool TimerContainer::AddTimer(std::weak_ptr<TimerSolt> t, uint32_t time, bool always) {
-    if (time >= _capacity) {
+    if (time >= (uint32_t)_capacity) {
         return false;
     }
 
-    if (time < _accuracy) {
+    if (time < (uint32_t)_accuracy) {
         return _sub_timer->AddTimer(t, time, always);
     }
 
@@ -48,7 +53,7 @@ bool TimerContainer::AddTimer(std::weak_ptr<TimerSolt> t, uint32_t time, bool al
 
     // relative to current time
     time += _cur_index * _accuracy;
-    if (time > _capacity) {
+    if (time > (uint32_t)_capacity) {
         time %= _capacity;
     }
 
@@ -123,7 +128,7 @@ uint32_t TimerContainer::TimerRun(uint32_t time) {
                 if (next_index >= _max_size) {
                     next_index = next_index % _max_size;
                 }
-                AddTimerByIndex(ptr, next_index);
+                AddTimerByIndex(ptr, uint8_t(next_index));
             }
         }
     }
